Refactor package structure to unify model naming conventions across the project
