+++ core/network/consumer-rules.pro
@@ -19,9 +19,6 @@
 # ... RETROFIT PROGUARD RULES ...
 # ... (https://github.com/square/retrofit/issues/3751)
 
-# Keep model classes used for deserialization by Retrofit.
--keep class dev.atick.core.network.models.** { *; }
-
 # Retrofit does reflection on generic parameters. InnerClasses is required to use Signature and
 # EnclosingMethod is required to use InnerClasses.
 -keepattributes Signature, InnerClasses, EnclosingMethod
@@ -63,3 +60,6 @@
 # kept. Suspend functions are wrapped in continuations where the type argument
 # is used.
 -keep,allowobfuscation,allowshrinking class kotlin.coroutines.Continuation
+
+# Keep model classes used for deserialization by Retrofit.
+-keep class dev.atick.core.network.model.** { *; }
