+++ core/preferences/consumer-rules.pro
@@ -1,2 +1,2 @@
 # Keep model classes used for deserialization.
--keep class dev.atick.core.preferences.models.** { *; }
+-keep class dev.atick.core.preferences.model.** { *; }
