+++ core/preferences/src/main/kotlin/dev/atick/core/preferences/model/DarkThemeConfigPreferences.kt
@@ -14,22 +14,23 @@
  * limitations under the License.
  */
 
-package dev.atick.core.network.models
+package dev.atick.core.preferences.model
 
+import dev.atick.core.preferences.model.DarkThemeConfigPreferences.DARK
+import dev.atick.core.preferences.model.DarkThemeConfigPreferences.FOLLOW_SYSTEM
+import dev.atick.core.preferences.model.DarkThemeConfigPreferences.LIGHT
 import kotlinx.serialization.Serializable
 
 /**
- * Data class representing a network post retrieved from a remote source.
+ * Enum class representing configuration options for the dark theme.
  *
- * @property id The unique identifier of the network post.
- * @property title The title of the network post.
- * @property url The URL associated with the network post.
- * @property thumbnailUrl The URL of the thumbnail image associated with the network post.
+ * @property FOLLOW_SYSTEM The dark theme configuration follows the system-wide setting.
+ * @property LIGHT The app's dark theme is disabled, using the light theme.
+ * @property DARK The app's dark theme is enabled, using the dark theme.
  */
 @Serializable
-data class NetworkPost(
-    val id: Int,
-    val title: String,
-    val url: String,
-    val thumbnailUrl: String,
-)
+enum class DarkThemeConfigPreferences {
+    FOLLOW_SYSTEM,
+    LIGHT,
+    DARK,
+}
