+++ data/src/main/kotlin/dev/atick/data/repository/settings/SettingsRepositoryImpl.kt
@@ -18,10 +18,10 @@ package dev.atick.data.repository.settings
 
 import dev.atick.core.preferences.data.UserPreferencesDataSource
 import dev.atick.core.utils.suspendRunCatching
-import dev.atick.data.models.settings.DarkThemeConfig
-import dev.atick.data.models.settings.Settings
-import dev.atick.data.models.settings.asDarkThemeConfigPreferences
-import dev.atick.data.models.settings.asSettings
+import dev.atick.data.model.settings.DarkThemeConfig
+import dev.atick.data.model.settings.Settings
+import dev.atick.data.model.settings.asDarkThemeConfigPreferences
+import dev.atick.data.model.settings.asSettings
 import dev.atick.firebase.auth.data.AuthDataSource
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
