+++ firebase/firestore/consumer-rules.pro
@@ -1,2 +1,2 @@
 # Keep model classes used for deserialization.
--keep class dev.atick.firebase.firestore.models.** { *; }
+-keep class dev.atick.firebase.firestore.model.** { *; }
